backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
websockets==15.0.1
pytest==8.4.1
google-adk>=1.26.0
google-genai>=1.66.0
google-cloud-aiplatform>=1.90.0
python-multipart>=0.0.20
